backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
pydantic==2.12.5
python-dotenv==1.0.1
google-genai>=1.55.0,<2
